--- exporter/pdf/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "typst-ts-pdf-exporter"
-description = "Export a frame of Typst document into a PDF document."
-version.workspace = true
-license.workspace = true
-edition.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-
-typst.workspace = true
-typst-pdf.workspace = true
-
-typst-ts-core.workspace = true
