--- exporter/pdf/README.md
@@ -1,3 +1,0 @@
-# typst-ts-pdf-exporter
-
-See [Typst.ts](https://github.com/Myriad-Dreamin/typst.ts)
