--- exporter/serde/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "typst-ts-serde-exporter"
-description = "Export a frame of Typst Artifact into serde formats."
-version.workspace = true
-license.workspace = true
-edition.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-
-typst.workspace = true
-
-serde.workspace = true
-serde_json = { workspace = true, optional = true }
-rmp-serde = { workspace = true, optional = true }
-
-typst-ts-core.workspace = true
-
-[features]
-json = ["serde_json"]
-rmp = ["rmp-serde"]
-default = []
