--- exporter/serde/README.md
@@ -1,3 +1,0 @@
-# typst-ts-serde-exporter
-
-See [Typst.ts](https://github.com/Myriad-Dreamin/typst.ts)
