--- exporter/serde/src/lib.rs
@@ -1,11 +1,0 @@
-pub(crate) use typst_ts_core::exporter_utils::*;
-
-#[cfg(feature = "json")]
-pub(crate) mod json;
-#[cfg(feature = "json")]
-pub use json::JsonExporter;
-
-#[cfg(feature = "rmp")]
-pub(crate) mod rmp;
-#[cfg(feature = "rmp")]
-pub use rmp::RmpExporter;
