+++ packages/compiler/Cargo.toml
@@ -41,14 +41,13 @@ web-sys = { workspace = true, features = [
 ] }
 console_error_panic_hook.workspace = true
 
-typst-ts-core.workspace = true
+typst-ts-core = { workspace = true, features = ["pdf"] }
 typst-ts-compiler = { version = "0.5.0-rc6", default-features = false, features = [
     "web",
 ] }
 typst-ts-ast-exporter.workspace = true
 reflexo-vec2canvas.workspace = true
 typst-ts-svg-exporter.workspace = true
-typst-ts-pdf-exporter.workspace = true
 
 [dev-dependencies]
 wasm-bindgen-test.workspace = true
@@ -61,5 +60,5 @@ web-sys = { workspace = true, features = ["console"] }
 web_test = ["typst-ts-test-common/web_artifacts"]
 
 full = ["web", "web_test"]
-web = []
+web = ["typst-ts-core/web"]
 default = ["full"]
