+++ tests/integration/Cargo.toml
@@ -24,12 +24,10 @@ tokio.workspace = true
 
 typst-ts-dev-server.workspace = true
 typst-ts-test-common.workspace = true
-typst-ts-core.workspace = true
+typst-ts-core = { workspace = true, features = ["pdf"] }
 typst-ts-compiler = { workspace = true, features = ["system"] }
 
-typst-ts-pdf-exporter = { workspace = true }
 typst-ts-svg-exporter = { workspace = true }
-typst-ts-serde-exporter = { workspace = true, features = ["json"] }
 
 [[test]]
 name = "tests"
