+++ tests/integration/src/lib.rs
@@ -5,13 +5,13 @@ use std::path::Path;
 use typst_ts_compiler::{
     CompileDriver, CompileExporter, PureCompiler, TypstSystemUniverse, TypstSystemWorld,
 };
+use typst_ts_core::PdfDocExporter;
 use typst_ts_core::{
     config::{entry::EntryOpts, CompileOpts},
     exporter_builtins::{FsPathExporter, GroupExporter},
     path::PathClean,
     TypstDocument,
 };
-use typst_ts_pdf_exporter::PdfDocExporter;
 use typst_ts_svg_exporter::SvgModuleExporter;
 
 fn get_driver(
